Remove external library headers, DLLs, LIBs
Up until now, this project has contained in-tree the headers and DLLs
for the libraries it uses. This works well on Windows, which doesn't
provide a system package manager. However, it makes things more
difficult on other platforms, which obviously can't use the included
DLLs; on those platforms, we must rely on an external source for
the libraries, and that can conflict with the in-tree headers.

Going forward, I'm going to remove all the in-tree libraries. We'll
rely on a package manager to supply the necessary headers and object
files outside of the tree.

Note: this will certainly break the build in Visual Studio. I'm
going to switch to a CMake-based build system, which will eventually
supplant the Visual Studio project.