GameObject::getModel(), getModelConst() returns pointer vs ref
GameObject instances can hold up to five models, each held by
pointer. When a model is not present (because there are less than
five models), the corresponding pointers are set to NULL.

Client code would iterate over the models by iterating over all five
model pointers, skipping those that were NULL. However, getModel()
and getModelConst() return references. It is not legal to return
null references. In debug mode, the program tends to work. However,
in optimized builds, the compiler assumes that (per the spec) the
returned value is never null; consequently, the code won't notice
when a model is not present and instead act as if the null reference
pointed to an actual model. This leads to segfaults.

This commit changes the two functions so that they return pointers
instead. It also updates all uses of the two functions to match.