Fix internal format used when generating textures in CTexture
Beginning with (it seems) OpenGL 3.0, the internalformat in a call to
glTexImage2D can't simply  be the number of channels in the texture (which,
in our case, would be 4: R, G, B, A). Instead, it has to be a proper
constant defined in the documentation. In our case, the right internal
format is GL_RGBA8. We used to simply specify 4, which was allowed by
OpenGL 2 and worked on some platforms but led to improperly-loaded
textures on, at the very least, OS X.