Use Apple-specific mouse warping to prevent event loss
At every frame, after reading the position of the mouse, we recenter it in
the frame. GLUT provides a function, glutWarpPointer(), to do this. However,
it does not work well on macOS. As it turns out, when warping the pointer
using that function, the Quartz event system suppresses mouse events for
around 0.25 seconds. That is not good.