Fix checking for empty string by direct literal comparsion
Not all empty strings are guaranteed to have the same address
(even between literals), so we can't simply compare a C-string
to another C-string using the == operator. Instead, we can simply
check to see if the first character is NULL. Note that we also
check if the pointer is NULL itself, which we also treat as an
empty string. (We have to check for pointer nullity before we
deference the pointer to check if the first character is null.)