Use SDL_TTF and textures to reliably render text on all computers
Prior to this commit, text was rendered using glut_bitmap_character.
On Macs (and any platforms that don't support compatibility profiles),
this function does nothing because core profiles don't support the
underlying OpenGL calls used by the GLUT function. We also had reports
of the text crashing the game on other computers, possibly related to
the same problem.

This commit changes the way we render fonts to use SDL_TTF to generate
a texture. We then borrow the UI hack from general.cpp's callbackDisplay(),
using a heavily distorted cube.obj to render the texture with a
orthographic matrix.

Note that SDL_TTF is a distinct dependency from SDL itself.