Replace GLUT with SDL2
Using GLUT on macOS has been difficult. FreeGLUT doesn't support
macOS without going through XQuartz, which presents its own problems
with OpenGL. macOS does provide a copy of GLUT, which we can use,
but it lacks some features that we use with FreeGLUT (namely exiting
the GLUT loop). It also necessitates some macOS-specific mouse handling
code.

Because we've already been using SDL2 for sound and text rendering,
it makes sense to simply switch over to SDL2 for video as well.
The switchover as implemented by this commit isn't ideal because
it strives to make the fewest changes to the code, which is already
organized around the structure of GLUT. But it works!