Allow player to rotate in fractional degrees
In order to slow down rotation to a reasonable rate, the Player
code divded the mouse delta by 10 to determine the number of degrees
to rotate for that frame. However, it would do integer division,
which meant that the player could never move a fractional degree ever.
The most problematic consequence of this was the slight mouse movements
would do nothing.

This commit simply changes the code so that it performs floating-point
divison, allowing for rotations of fractional degrees.