Change into executable's directory at start
The resources directory is supposed to be in the same directory as the
binary. If the working directory isn't the binary's directory, we won't be
able to find any of the resources.

This commit attempts to determine the binary's parent directory and changes
into it at the start of main().