Fix SDL_Surface width discrepancy in Text2D on new Mac
On some platforms (e.g., my new M2 Mac running Ventura), the length of each
row in bytes is not equal to the number of pixels per row times the number
of bytes per pixels. This might have something to do with alignment or other
platform optimization concerns. The surface's pitch is the number of bytes
per row; we can use it to calculate the size of each row in pixels and
provide it to OpenGL.