Add support for CMAKE_INSTALL_PREFIX
If CMAKE_INSTALL_PREFIX is explicitly set, the install target will be
configured so that the binary is installed in the bindir and the
resourced directory installed in the datadir (e.g., "share").
Furthermore, the datadir will be provided as a preprocessor defintion;
main() will chdir into it so that the resources directory can be found.

Note that, if CMAKE_INSTALL_PREFIX is not explicitly set (and remains
the default value), the old behavior is kept. That is, the resources
directory is copied into the build directory and main() attempts to
find the parent of the executable (i.e., the build directory) and
changes into it. This way, the program can be executed straight out
of the build directory.