Prevent canvas collapsing to 0 width on return from fullscreen in Safari
This regression was caused by turning the emscripten loading panel into
an overlay. As part of that, div#video lost its "width: 100%" style.

It's not as simple as adding it back in, though, because the new layout
would cause the center column (when flowed horizontally) to expand to
take up all slack. Instead, both div#video and div#emscripten-border
have their min-width set to `fit-content`. This allows it to grow to
the viewport's full width when the flex container is flowed vertically,
which, in turn, allows the canvas to be centered via its auto left and
right margins.