Center overlay over canvas, even if flexbox is taller
Before this commit, div#video would be as tall as the flexbox,
whose height is tha max of its children. Normally, that's the
canvas, so the overlay would be properly centered it in. However,
if either side panel were to be longer than the canvas, the overlay
would end up not being centered over the canvas.

This is fixed by limiting the height of div#video to fit-content.