Start tick timer before invoking callbackTimer on Emscripten
Before this commit, the next tick would execute FRAME_INTERVAL_MS
milliseconds after the callbackTimer returned. After this commit,
the next tick would execute the same number of milliseconds after
the start of the current tick.

This new way is basically the way things work on native platforms
using the SDL timer. It'll make things easier in the future if
both platforms behave the same way.