Add support for gamepads
This implementation contains a bug: if the player, running on the
Emscripten platform, starts the game using the gamepad's start
button, the background music will not start playing.

There's also an oddity here that stems from the way this support
is implemented: by faking keystrokes and mouse activity. The player
can start a game with the start button, the right trigger, and the
Y button (which normally activates the shield). They cannot do so,
however, by jumping or scrolling through weapons, because these
are keyboard actions and not mouse clicks.

Finally, note that this always uses the first enumerated or connected
gamepad. There's no way to select the gamepad.