Add favicon.png [em]
This is the first "web resource." That is, it is delivered alongside
the other web files (e.g., the HTML and WASM files). We use the same
technique, more or less, as delivering the runtime resources for the
native application.

If the installation prefix is not set, the icon is symbolically
linked into the build directory so it can be accessed when the app
is served from there. (Notably, this is different from the native
build, which copies the files instead of linking them.)

If the installation prefix is set, however, the icon is installed
alongside the WASM and other files instead.