Add/use install target [em]
Instead of having custom code in the flake to copy files from the build
directory and then largely duplicative code in emscripten-webpage.yml,
this commit changes things by adding an install target to CMakeLists.txt.
The flake delegates entirely to the target; the GitHub action just copies
all the files (with the HTML file renamed to "index.html").

This has the benefit of centralizing the declaration of all runtime
artifacts in the CMakeLists.txt file. This isn't that useful now, but it
will make it a lot easier if we want to ship/deploy additional files
(e.g., the favicon or the intro screen as a separate file).

---

As an aside, the installation target for the Emscripten build is better
behaved compared to its native counterpart in that it installs into a
project-specific subdirectory in the share folder. This is pretty
useless because it's highly unlikely that an emscripten build will be
installed into a user's profile, but it's the more correct way to do it.
The native install target should probably be updated to follow suit.