Fix left padding of HUD depending on aspect ratio
As the window becomes increasingly wider, the HUD moves more to the right.
As the window becomes increasingly narrower, it moves more to the left.
This doesn't really make much sense. This commit fixes it so the HUD is
properly anchored to the left edge of the window.

This commit fixes a problem somewhat similar to that fixed in bc20d524.
Before that commit, there was a gap between the health/shield bars and
their labels that would vary depending on the window's aspect ratio.