Back out "Don't play audio until animation starts"
Original commit changeset: e54173171d2f

Now that we display the intro screen on Emscripten builds and delay
starting the program until the user activates it, we don't need to
delay the background music until animation starts. That is, we now
unconditionally start the game animated, so we don't need to support
holding off on playing the music.