Suppress mouse motion events from first 2 frames [em]
It seems like, on Firefox and Safari, we get an early mouse motion
event within the first two frames before the mouse is centered.
As a result, the tank effectively starts looking in the wrong
direction.

This diff suppresses mouse motion events for the first two frames.
Technically, the counter overflows in around 54 months, but that's
not a big deal if it somehow does because the counter overflows to
zero.