Add (un)pause() and friend functions, (un)pause audio immediately
Before this commit, background audio was synced with the pause
state every timerCallback(). However, this poses a problem on the
web; when the window isn't visible, the timerCallback(), which is
connected to requestAnimationFrame(), won't get called.

This commit changes pauseAnimation so that its a read-only view
of the pause state. Mutation must go through one of the following
new functions:
  - pause()
  - unpause()
  - setPause(bool)
  - togglePause()

When the pause state changes, the background music state changes
with it. This means that it'll have an immediate effect and no
longer depends on an animation frame callback being invoked.

There is a consequence to this: calling SoundPlayer::playBackground()
will start music immediately, even if the game is paused. For
example, if we wanted to start the game paused, we'd have to do
one of the following:
  - immediately pause the music after starting it if the game is
    paused
  - pause the game after starting the music

It would also be possible to either tell SoundPlayer about the
pauseAnimation flag (which seems leaky) or add code to SoundPlayer
to have its own pause state which starts music paused when asserted.