Pause game when window lacks focus [em]
The main benefit of this is on mobile devices, where the background
music will be paused when the user changes tabs or switches apps.

(Note that, because the game's loop is driven by requestAnimationFrame()
and that callback is generally not invoked when the window/tab is hidden,
the game was already effectively paused when the window was hidden. That
is distinct from losing focus, though, which is a broader category.)