Clamp Text2D textures to prevent edge wrapping artifacts
Probably due to min/mag sampling, sometimes the edge of one side of text
will appear on the opposite side. This commit prevents this from happening
by clamping the texture coordinates and then properly setting up the S/T
offsets so that the camera-facing side lines up with the texture.