Explictly list preloaded resource files [em]
It turns out the the resources directory contains an extra 8.17 MB
of data that we don't use. To prevent us from bundling and delivering
them to client, we now list precisely the bundled resoures that we
need and include only those.

This also allows us to remove the GLOB operation to determine the
link dependencies for the target; we can just use the specifically
enumerated resource files instead.