Deliver hud, intro-screen images as web resources [em]
Before this commit, they were bundled in the .data file and needed
to be exposed through blob URIs before being used. This makes the
HTML/CSS simpler; because the URIs are known ahead of time, we can
use static CSS and a simple class to turn the intro screen on and
off.

Furthermore, this lets us put the code to turn off the intro screen
in startGame() (which is defined before the blob URIs are known)
easily. Importantly, this actually has a small behavioral change:
the intro screen is removed if main() runs successfully, even if
the subsequent code in startGame() (i.e., to deal with a fullscreen)
throws.