Use javascript to center 404 error message when wrapped [em]
The intrinsic width of a block doesn't change when its text wraps. As a
result, it's not possible to center a block containing margin-aligned,
wrapped text; the width of the block will get stuck at the size of its
container, which will effectively force the text to be flush with one
edge.

Prior to this commit, we handled this by using a container query with
a ch-denominated minimum width (taking advantage of the fact that the
font is monospace and that we can be approximate here) to trigger a
set of rules when the text is about to wrap (with some padding, for
aesthetics). This set forces the text box to its minimum size, which
the box's intrinsic width will match.

However, this produces overly wrapped text. To fix this, this commit
uses Javascript and some additional spans nest in each paragraph.
Whenenver the page is resized, we measure the width of each span and
take the largest. We then compare it to the width of the box. If they
are equal, the box is centered (either because there's no wrapping or
because the wrapping lined up perfectly with the right edge). Otherwise,
we need to perform an adjustment: the box's X position is transformed
by an amount equal to half the difference.

CSS transformations affect rendering but not layout! This is important!
We can't simply move the box to the right to compensate; that'll just
make the box smaller and induce more wrapping. We also can't adjust the
box's width to fit exactly its contents; from that point, the width
would be fixed and we'd be unable to detect when we should re-enlarge it.