Optionally compress textures to webp [em]
This commit adds USE_WEBP_TEXTURES, a CMake cache variable that defaults to
OFF. When set to ON, the bundled jpgs and pngs are first compressed to webp
using the cwebp utility. It also defines a preprocessor definition of the
same name, which will trigger CTexture to append ".webp" to all texture
filenames.

Furthermore, it updates the emscripten package in the flake to turn on this
functionality.

The reason for making this optional is the dependency on cwebp (provided by
the libwebp dependency). Requiring the dependency would potentially make it
more complicated to set up Visual Studio Code, especially on a GitHub
Codespace.

This change saves 1.2 MiB from the build output.