Invalidate WorldEntity translate cache for all when any is altered
As noted in the comment present before this commit,
the return value of WorldEntity::getTranslate() is cached for each entity
and reset every frame. This means that the returned value may be stale.
As it turns out, this was the probable source of some bugs where the player
or other moving entities would warp through or around objects. Changes to
the entity's position as a result of handling a collision would not be
reflected in the entity's getTranslate() value, which would cause further
collisions processed in the same frame to operate on bad positions.

WorldEntity instances do not have pointers back to their children, so we
cannot invalidate specifically those cached values that might be affected
by a change to an entity's relative translation, rotation, or scale. As a
result, this commit invalidates the entire cache for *all* objects whenever
any single object's local translation, rotation, or scale is changed.

However, as an optimization, this commit also adds a flag to WorldEntity,
_isParent, which is set whenever an entity is parented to it. This flag
is never cleared. (So, the flag more accurately represents if the object
is *or ever was* a parent.) If this flag is false, any change to the
object's transformation state is isolated, so we can just invalidate its
own cache. This is important in the case of temporary objects (like when
rendering hitboxes), which would otherwise blow away the cache when setting
their initial transformation state via calls to setTranslate and the like.