Install DLL dependencies alongside EXE on Windows
Unlike on MacOS and Linux, where executable files contain embedded
paths to their librares (which are typically installed in a common
location), Windows generally expects an executable to come with its
DLLs. This commit causes the installation process to copy the app's
DLLs (and their transitive dependencies) to the bin directory.

This is somewhat complicated. CMake provides two mechanisms for
figuring out an executable's transitive dependencies, but they
have pitfalls:
1) $<TARGET_RUNTIME_DLLS:...>. This seems more modern, but it only
   captures some and not all of the executable's *transitive*
   dependencies.
2) FILE(GET_RUNTIME_DEPENDENCIES ...) and the higher-level
   runtime dependency set. When applied to the executable, this
   detects many system DLLs. While the ones in system32 can be easily
   filtered out, it also generates a lot of unresolved dependencies
   for other system DLLs.

This commit takes a hybrid approach:
1) It installs the DLLs found via the $<TARGET_RUNTIME_DLLS:...>
   mechanism.
2) It then uses custom installation code that calls the lower-level
   file(GET_RUNTIME_DEPENDENCIES ...) mechanism on the DLLs from
   step 1 to get their recursive dependencies, filtering out all
   DLLs in system32. This doesn't detect the other system DLLs that
   failed to resolve. It produces a clean list of DLLs that can be
   copied over.