Mark executable as WIN32
This shouldn't do anything on other platforms, but it's necessary on
Windows to indicate that the application is graphical. Its main
(maybe even only) consequence is to prevent a command prompt from
launching when starting the application.

It does have a disadvantage, though: stdout and stderr will no longer
appear when running the application. They can be captured, though,
via file redirects with cmd.exe (but, notably, not PowerShell).