Try to attach parent console on Windows
On Windows, we compile as a graphical application to suppress the creating
of a terminal window when launching it. However, if the app was started via
a terminal, this prevents stdout/stderr from reaching it. Here, if we were
started in a terminal, we'll attach the app to its console so we can write
to it.